+++ src/errors/error_runtime.rs
@@ -61,8 +61,8 @@ pub enum RuntimeErrorKind {
     MethodNotFound(Cow<'static, str>, Cow<'static, str>),
     // #[error("module `{0}` not found")]
     // ModuleNotFound(Cow<'static, str>),
-    #[error("no module defined for `{0}`:{1}")]
-    NoModuleDefined(String, Cow<'static, str>),
+    #[error("no module defined for {1} during {2}:\n `{0}`")]
+    NoModuleDefined(String, Cow<'static, str>, Cow<'static, str>),
     #[error("not a callable function: `{0}`")]
     NotAFunction(String),
     #[error("type error, expected `{expected}`, found `{found}`:\n  {sym}")]